+++ packages/semantic-ui/src/Theme/Theme.js
@@ -17,7 +17,7 @@ const Theme = {
   fields: { ...fields, ...Fields },
   FieldTemplate,
   ObjectFieldTemplate,
-  tagName: SuiForm,
+  _internalFormWrapper: SuiForm,
   widgets: { ...widgets, ...Widgets },
   children: React.createElement(SubmitButton)
 };