+++ CHANGELOG.md
@@ -15,7 +15,24 @@ it according to semantic versioning. For example, if your PR adds a breaking cha
 should change the heading of the (upcoming) version to include a major version bump.
 
 -->
-# v4.0.1 (upcoming)
+# v5.0.0 (coming soon)
+
+# v4.0.2 (upcoming)
+
+# v4.0.1
+
+- Bumped the peer dependencies of `@rjsf/core` to `^4.0.0` for all of themes in `package.json`
+- Also, added tests to all themes to verify that the `tagName` prop works as expected
+
+## @rjsf/core
+- Updated `Form` to support the `semantic-ui` and `material-ui` themes to allow them work when `tagName` is provided
+
+## @rjsf/material-ui
+- Fixed up the `Theme` and `Theme5` implementations to deal with a regression in which adding `tagName` caused the 2 themes to not work
+- Only `console.log()` the import error in `MaterialUIContext` and `Mui5Context` when not in `production` to eliminate tons of warnings for released code
+
+## @rjsf/semantic-ui
+- Fixed up the `Theme` implementation to deal with a regression in which adding `tagName` caused the theme to not work
 
 # v4.0.0